Refactor: move form related components and functionality to subfolder
