+++ src/app/components/form/oph-input.tsx
@@ -1,0 +1,17 @@
+'use client';
+import React from 'react';
+import { OutlinedInput, InputProps } from '@mui/material';
+
+export const OphInput = ({
+  value,
+  onChange,
+  ...props
+}: InputProps & {
+  value: string;
+  onChange: (event: React.ChangeEvent<HTMLInputElement>) => void;
+  helperText?: string[];
+}) => {
+  return (
+    <OutlinedInput value={value} size="small" onChange={onChange} {...props} />
+  );
+};
